test: add integration test for `ReqwestClient`
