--- sec/tests/integration_tests.rs
@@ -1,1 +1,0 @@
-
